61) What can you do to spill the reducers? We can use event actions.
62) How do you implement routing in ReactJS?
React Router is available as an npm package, and can be added to your application with a single line of code. Once installed, you can define your routes in your application and use the component to link between them.

63) Can you attach a JSX element to other JSX components?
Yes, you can add JSX elements to other JSX components.

64) How can you fetch data with React Hooks?
We can use useEffect that fetches the data through axios from the API.

65) What is one-way data flow in React?
React uses unidirectional data flow which means you can transfer data from parent to child and not the other way around.

66) Why do we use keys in Lists?
The main reason why we use keys in lists are as follows
It is an identifier to find which items have changed, updated, or deleted from lists.